Add dependency, documents and linting.
